Add error handling
Error handling allows a dependent to check for errors nicely.
You don't have to check for undefined anymore.

Just use the err variable transmitted with the init-callback.

The error either transmits the err from the read call or builds an own error.
The own error simply states that no packages weren't found.

If everything is fine, err will be null,
For downward compatibility to other dependents, the err is submitted after the sorted (or filtered)