FEATURE: Improve license exclusion
The ability to exclude licenses is great. This spruces it up a little.

Valid SPDX identifiers are combined into an SPDX expression for more
accurate checking.
Invalid SPDX identifiers are checked via simple string comparison.

In addition, BSD is aliased to an expression that includes all the basic
BSD-n-Clause licenses.