Fixes #177, --packages and --excludePackages filtering
The filtering logic for both `--packages` and `--excludePackages` was broken if the output was being colorized.

To work around this I have moved the colorization out of the parsing logic and into the formatting.

This change means we no longer need to use `strip-colors` so I've removed this as a dependency.