replace `process.exit()` with throwing an error
Using process.exit() does not allow us to use this tool programmatically, as we cannot then handle the error in a custom way (ie: formatting the response for CI, logging it, somewhere, etc). 

This PR replaces logging to `console.error` + using `process.exit` with throwing an error.  This will have the same effect in CLI usage by immediately terminating the process with a message, but can also be `try/catch` handled in a parent program.